add the python3 shebang
The shebang line in any script determines the script's ability be executed like a standalone executable without typing python beforehand in the terminal or when double-clicking it in a file manager (when configured properly). It isn't necessary but generally put there so that when someone sees the file opened in an editor, they immediately know what they're looking at. However, which shebang line you use is important.